Changed checkbox input handling
Previously when encoutering checkbox input script filled all possible answers in given quetstion. As this is propably not such random, i changed that to choose just one of possible values. The best solution would be to choose from 1 to n values from possible input values and then send that, but i'll work on it other day